# docs/swagger.yaml
basePath: /
definitions:
  models.Recipe:
    properties:
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      instructions:
        items:
          type: string
        type: array
      name:
        type: string
      publishedAt:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  models.User:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is an Recipe API
  termsOfService: http://swagger.io/terms/
  title: Recipe API
  version: "1.0"
paths:
  /recipes:
    get:
      description: Lists all the recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
      summary: Lists all the Recipes in the db
      tags:
      - recipes
    post:
      description: This endpoint creates a new recipe in the db
      parameters:
      - description: The new Recipe
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Creates a new Recipe
      tags:
      - recipes
  /recipes/{id}:
    delete:
      description: Deletes a recipe from the db given an id
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted Successfully
          schema:
            additionalProperties: true
            type: object
      summary: Deletes a recipe
      tags:
      - recipes
    get:
      description: Show a specific recipe
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe not found
          schema:
            additionalProperties: true
            type: object
      summary: Show a recipe
      tags:
      - recipes
    put:
      description: Update an existing recipe given an ID parameter
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated recipe data
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe not found
          schema:
            additionalProperties: true
            type: object
      summary: Update a recipe
      tags:
      - recipes
  /recipes/search:
    get:
      description: This endpoint will search the db and return the recipes within
        that tag.
      parameters:
      - description: Tag to search recipes
        in: query
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Recipe not found
          schema:
            additionalProperties: true
            type: object
      summary: Search a recipe from a tag
      tags:
      - recipes
  /refresh:
    post:
      description: This endpoint lets a user refresh their token cridential
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Refresh token
      tags:
      - auth
  /signin:
    post:
      description: This endpoint lets a user login with a username and a password
      parameters:
      - description: Updated recipe data
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Login with username and password
      tags:
      - auth
  /signout:
    post:
      description: This endpoint lets a user signout from the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Signout user
      tags:
      - auth
swagger: "2.0"