# handlers/auth.go
package handlers

import (
	"context"
	"crypto/sha256"
	"encoding/hex"
	"net/http"
	"time"

	"github.com/dgrijalva/jwt-go"
	"github.com/gin-contrib/sessions"
	"github.com/gin-gonic/gin"
	"github.com/natnael-alemayehu/recipes-api/models"
	"github.com/rs/xid"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/mongo"
)

type AuthHandler struct {
	collection *mongo.Collection
	ctx        context.Context
}

func NewAuthHandler(ctx context.Context, collection *mongo.Collection) *AuthHandler {
	return &AuthHandler{
		collection: collection,
		ctx:        ctx,
	}
}

type Claims struct {
	Username string `json:"username"`
	jwt.StandardClaims
}

type JWTOutput struct {
	Token   string    `json:"token"`
	Expires time.Time `json:"expires"`
}

// SigninHandler godoc
//
//	@Summary		Login with username and password
//	@Description	This endpoint lets a user login with a username and a password
//	@Produce		json
//	@Tags			auth
//	@Success		200	{object}	string
//	@Param			recipe	body		models.User	true	"Updated recipe data"
//	@Failure		400 {object}	map[string]interface{} "error"
//	@Router			/signin	[post]
func (handler *AuthHandler) SignInHandler(c *gin.Context) {

	var user models.User

	if err := c.ShouldBindJSON(&user); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error(),
		})
		return
	}
	h := sha256.New()
	cur := handler.collection.FindOne(handler.ctx, bson.M{
		"username": user.Username,
		"password": hex.EncodeToString(h.Sum([]byte(user.Password))),
	})
	if cur.Err() != nil {
		c.JSON(http.StatusUnauthorized, gin.H{
			"error": "Invalid username or password",
		})
		return
	}

	sessionToken := xid.New().String()
	session := sessions.Default(c)
	session.Set("username", user.Username)
	session.Set("token", sessionToken)
	session.Save()

	c.JSON(http.StatusOK, gin.H{"mesasge": "User signed in"})
}

// RefreshHandler godoc
//
//	@Summary		Refresh token
//	@Description	This endpoint lets a user refresh their token cridential
//	@Produce		json
//	@Tags			auth
//	@Success		200	{object}	string
//	@Failure		400 {object}	map[string]interface{} "error"
//	@Router			/refresh [post]
func (handler *AuthHandler) RefreshHandler(c *gin.Context) {
	session := sessions.Default(c)
	sessionToken := session.Get("token")
	sessionUser := session.Get("username")
	if sessionToken == nil {
		c.JSON(http.StatusUnauthorized, gin.H{
			"Error": "Invalid session cookie",
		})
		return
	}

	sessionToken = xid.New().String()
	session.Set("username", sessionUser.(string))
	session.Set("token", sessionToken)
	session.Save()

	c.JSON(http.StatusOK, gin.H{"messge": "new session issued"})
}

// RefreshHandler godoc
//
//	@Summary		Signout user
//	@Description	This endpoint lets a user signout from the system
//	@Produce		json
//	@Tags			auth
//	@Success		200	{object}	string
//	@Failure		400 {object}	map[string]interface{} "error"
//	@Router			/signout [post]
func (handler *AuthHandler) SignoutHandher(c *gin.Context) {
	session := sessions.Default(c)
	session.Clear()
	session.Save()

	c.JSON(http.StatusOK, gin.H{"mesage": "Signed out..."})
}